Correct workflow artifact name pattern in size deltas report workflow The "sketches-reports-source" input of the "arduino/report-size-deltas" GitHub Actions action defines the regular expression that matches the names of the sketches report workflow artifacts produced by the "Compile Examples" workflow.
The key string in the names of these artifacts was set to "sketches-report" when the "Compile Examples" workflow was
adjusted for compatibility with the breaking changes introduced by updating to version 4.x of the workflow's
"actions/upload-artifact" GitHub Actions action dependency. The pattern set in the size deltas report workflow was
"sketches-reports". The "s" at the end of that pattern caused it to no longer match against the key string in the
artifact names after that adjustment of the "Compile Examples" workflow, resulting in size deltas reports no longer
being generated by the workflow.

Although a minimal fix would be to simply remove the "s" from the end of the pattern, the decision was made to use a
more strict regular expression. This will make it easier for maintainers and contributors to understand that this value
is a regular expression and the exact nature of how that regular expression functions (which is less clear when relying
on the "arduino/report-size-deltas" action's partial pattern matching behavior).