Add dedicated license file
Standardization in license documentation is important because, in addition to making it easy for humans to find this
vital information, it allows machines to automate the process of license type determination, which is useful both for
discovering suitable open source projects as well as checking open source license compliance.

Previously, the open source license of this project was not documented in a dedicated license file. Due to this, even
though the project is licensed under a standardized open source license, the open license type could not be identified
with 100% confidence by machines (e.g., the Licensee Gem used by the GitHub website), which meant identification could
only be made by a human carefully evaluating the license text.

The project's license is hereby documented in a standardized license file. The result is that the project's open source
license type can now be automatically identified. This does not result in any change to the project licensing.